refactor(scene): extract wandering-creature behaviour into crate::creatures
drawable.rs (1980 LOC) mixed the y-sort painter's-algorithm with two full
creature behaviour engines — the office pet and the OpenClaw gateway mascot.
That's a sim/render layering violation: WHERE a creature roams each frame is
behaviour (sim), not appearance (render), yet it lived in the pixel painter.

Move pet_position + mascot_* (+ their deliberately-shared roaming toolkit:
visit-spot geometry, walk_between, sample_polyline) + their behaviour tests
into a new scene-level module crate::creatures. The PAINT stays in
pixel_painter (paint_drawable's Pet/GatewayMascot arms + paint_mascot_bubbles)
— "scene decides, painter draws". pixel_painter/mod.rs now calls
crate::creatures::{pet_position, mascot_position, gateway_mascot_def}.

Pure code motion: byte-identical fn bodies, entry points pub(super)→pub(crate),
epoch_ms a local one-line forwarder to crate::anim::elapsed_ms (== the
pixel_painter one). drawable.rs 1980→1211, new creatures.rs 799. Full scene
suite green (627 tests), zero pixel change.

Kept pet + mascot in ONE creatures.rs (not two files) because they SHARE the
roaming toolkit; split by entity when a 2nd pet/mascot lands (YAGNI).
